pyproject.toml
[tool.poetry]
name = "trading-bot"
version = "0.1.0"
description = "A stock trading bot using Alpaca and Twilio"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
yfinance = "^0.1.70"
pandas = ">=1.1.0, <2.0.0"
numpy = ">=1.19.0, <1.22.0"
requests = "^2.26.0"
python-dotenv = "^0.19.0"
twilio = "^6.60.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
